+++ wled00/wled.h
@@ -742,7 +742,7 @@ WLED_GLOBAL bool notifyHue    _INIT(false);                       // send notifi
 #endif
 
 // effects
-WLED_GLOBAL byte effectCurrent _INIT(0);
+WLED_GLOBAL uint16_t effectCurrent _INIT(0);
 WLED_GLOBAL byte effectSpeed _INIT(128);
 WLED_GLOBAL byte effectIntensity _INIT(128);
 WLED_GLOBAL byte effectPalette _INIT(0);
